fix(web): #10
