+++ packages/babel-preset-rsmax/package.json
@@ -24,6 +24,7 @@
     "@babel/preset-env": "^7.24.7",
     "@babel/preset-react": "^7.24.7",
     "@rsmax/babel-plugin-auto-css-modules": "workspace:*",
+    "@babel/preset-typescript": "^7.24.7",
     "babel-plugin-auto-import": "^1.0.5",
     "babel-plugin-macros": "^3.1.0",
     "lodash": "^4.17.15"
