+++ packages/babel-preset-rsmax/src/index.ts
@@ -2,6 +2,9 @@ import { declare } from '@babel/helper-plugin-utils';
 
 interface PresetOption {
   react?: boolean | { [key: string]: any };
+  typescript?: any;
+  decorators?: any;
+  'class-properties'?: any;
   'throw-if-namespace'?: boolean;
   target?: any;
 }
@@ -10,6 +13,7 @@ function preset(api: any, presetOption: PresetOption) {
   api.assertVersion(7);
 
   const react = typeof presetOption.react === 'undefined' ? true : presetOption.react;
+  const typescript = typeof presetOption.typescript === 'undefined' ? true : presetOption.typescript;
   const throwIfNamespace =
     typeof presetOption['throw-if-namespace'] === 'undefined' ? false : presetOption['throw-if-namespace'];
   const targets =
@@ -18,6 +22,10 @@ function preset(api: any, presetOption: PresetOption) {
       : presetOption.target;
 
   const presets: any[] = [[require.resolve('@babel/preset-env'), { targets }]];
+
+  if (typescript) {
+    presets.push([require.resolve('@babel/preset-typescript'), typeof typescript === 'object' ? typescript : {}]);
+  }
 
   if (react) {
     const defaultReactOpt = { throwIfNamespace, runtime: 'automatic' };
