+++ packages/rsmax-cli/package.json
@@ -27,9 +27,6 @@
     "@babel/core": "^7.7.4",
     "@babel/helper-module-imports": "^7.7.4",
     "@babel/parser": "^7.7.4",
-    "@babel/preset-env": "^7.7.4",
-    "@babel/preset-typescript": "^7.7.4",
-    "@babel/register": "^7.7.0",
     "@babel/types": "^7.7.4",
     "@rsdoctor/cli": "^1.1.2",
     "@rsdoctor/rspack-plugin": "^1.1.2",
@@ -44,6 +41,7 @@
     "@rsmax/web": "workspace:*",
     "@rspack/core": "^1.3.12",
     "@rspack/dev-server": "^1.1.2",
+    "@swc-node/register": "^1.10.10",
     "babel-loader": "^9.1.3",
     "babel-plugin-module-resolver": "^5.0.2",
     "babel-plugin-rsmax-host-component": "workspace:*",
