+++ packages/rsmax/web.d.ts
@@ -1,0 +1,1 @@
+export * from '@rsmax/web';
