+++ packages/rsmax/web.js
@@ -1,0 +1,1 @@
+export * from '@rsmax/web';
